third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-in-meta-disallowed-console-warning.html

<!DOCTYPE html>
<html>
    <head>
        <script>
        if (window.testRunner)
            testRunner.dumpAsText();
        </script>
        <meta http-equiv="Suborigin" content="foobar">
    </head>
</html>
