third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-in-meta-disallowed-console-warning-expected.txt

CONSOLE ERROR: line 8: Error with Suborigin header: Suborigin header with value 'foobar' was delivered via a <meta> element and not an HTTP header, which is disallowed. The Suborigin has been ignored.
